refactor: algorithm directory structure (#105)
* put each algorithm in separate file
* consistently use underscores in algorithm directory names
