+++ examples/transport/main.cpp
@@ -1,10 +1,13 @@
 #include <fmt/core.h>
-#include <graaflib/algorithm/graph_traversal.h>
-#include <graaflib/algorithm/shortest_path.h>
+#include <graaflib/algorithm/graph_traversal/breadth_first_search.h>
+#include <graaflib/algorithm/graph_traversal/common.h>
+#include <graaflib/algorithm/shortest_path/bfs_shortest_path.h>
+#include <graaflib/algorithm/shortest_path/dijkstra_shortest_path.h>
 #include <graaflib/graph.h>
 #include <graaflib/io/dot.h>
 #include <graaflib/types.h>
 
+#include <optional>
 #include <unordered_set>
 
 struct station {
