+++ include/graaflib/algorithm/cycle_detection/dfs_cycle_detection.tpp
