+++ include/graaflib/algorithm/minimum_spanning_tree/prim.h
@@ -7,18 +7,6 @@
 #include <vector>
 
 namespace graaf::algorithm {
-/**
- * Computes the minimum spanning tree (MST) or minimum spanning forest of a
- * graph using Kruskal's algorithm.
- *
- * @tparam V The vertex type of the graph.
- * @tparam E The edge type of the graph.
- * @param graph The input graph.
- * @return A vector of edges forming the MST or minimum spanning forest.
- */
-template <typename V, typename E>
-[[nodiscard]] std::vector<edge_id_t> kruskal_minimum_spanning_tree(
-    const graph<V, E, graph_type::UNDIRECTED>& graph);
 
 /**
  * Computes the minimum spanning tree (MST) of a graph using Prim's algorithm.
@@ -37,4 +25,4 @@ template <typename V, typename E>
 
 }  // namespace graaf::algorithm
 
-#include "minimum_spanning_tree.tpp"
+#include "prim.tpp"
