+++ test/graaflib/algorithm/coloring/greedy_coloring_test.cpp
